90's Initial snake game
The fantasy snake game project is composed in python. This is a simple oop and GUI based fantasy game which is easy to understand and play. 
The trendy game of the 90's.
Speaking about the game play , all the polices are same as the actual version. The players get points when the snake eats up the food and grows  in size.
like., you have to eat more and stay in the game for as long as possible.